move in_memory module to its own file
